abstract away some of the esri time fetching and adding to map logic
